refactor: make some template modules private
