+++ src/template/aoc_cli.rs
@@ -11,7 +11,6 @@ pub enum AocCommandError {
     CommandNotFound,
     CommandNotCallable,
     BadExitStatus(Output),
-    IoError,
 }
 
 impl Display for AocCommandError {
@@ -22,7 +21,6 @@ impl Display for AocCommandError {
             AocCommandError::BadExitStatus(_) => {
                 write!(f, "aoc-cli exited with a non-zero status.")
             }
-            AocCommandError::IoError => write!(f, "could not write output files to file system."),
         }
     }
 }
